requirements.txt
asgiref==3.6.0
astor==0.8.1
attrs==23.1.0
bandit==1.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
darglint==1.8.1
Django==4.2
django-filter==23.2
djangorestframework==3.14.0
docutils==0.20.1
eradicate==2.2.0
flake8==3.9.0
flake8-bandit==2.1.2
flake8-broken-line==0.3.0
flake8-bugbear==21.11.29
flake8-commas==2.1.0
flake8-comprehensions==3.12.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.4.0
flake8-isort==4.2.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.2
flake8-rst-docstrings==0.2.7
flake8-string-format==0.3.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
isort==5.12.0
Markdown==3.4.3
mccabe==0.6.1
pbr==5.11.1
pep8-naming==0.11.1
psycopg2-binary==2.9.6
pycodestyle==2.7.0
pydantic==1.10.7
pydocstyle==6.3.0
pyflakes==2.3.1
Pygments==2.15.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
restructuredtext-lint==1.4.0
smmap==5.0.0
snowballstemmer==2.2.0
sqlparse==0.4.4
stevedore==5.1.0
typing_extensions==4.4.0
urllib3==2.0.2
wemake-python-styleguide==0.15.3
